requirements.txt
streamlit
huggingface_hub
boto3
Pillow
python-dotenv
